refactor(EditCollective): Move files to own folder
