+++ components/edit-collective/EditVirtualCards.js
@@ -10,15 +10,15 @@ import memoizeOne from 'memoize-one';
 
 import { Add } from '@styled-icons/material/Add';
 
-import { Router } from '../server/pages';
-import StyledSelect from './StyledSelect';
-import VirtualCardDetails from './VirtualCardDetails';
-import Loading from './Loading';
-import Pagination from './Pagination';
-import Link from './Link';
-import StyledButtonSet from './StyledButtonSet';
-import { P } from './Text';
-import StyledButton from './StyledButton';
+import { Router } from '../../server/pages';
+import StyledSelect from '../StyledSelect';
+import VirtualCardDetails from '../VirtualCardDetails';
+import Loading from '../Loading';
+import Pagination from '../Pagination';
+import Link from '../Link';
+import StyledButtonSet from '../StyledButtonSet';
+import { P } from '../Text';
+import StyledButton from '../StyledButton';
 
 const messages = defineMessages({
   notBatched: {
